requirements.txt
# Core ML/LLM Dependencies
faiss-cpu
sentence-transformers
huggingface-hub>=0.20.0
ctransformers==0.2.27

# GUI and API
gradio==4.21.0
pydantic==2.5.0
pydantic-core==2.14.1
fastapi==0.104.1
starlette==0.27.0

# Data Processing
tqdm
lxml
beautifulsoup4
pytest
requests
feedparser
scispacy
spacy>=3.7,<3.8

# Graph Database and GraphRAG
neo4j==5.19.0
neo4j-graphrag>=1.7.0  # Official Neo4j GraphRAG package for Python

# LangChain
langchain==0.1.16
langchain-community==0.0.32

# Core ML Frameworks for M1 Training
torch>=2.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
safetensors>=0.4.0
einops>=0.7.0

# Apple Silicon Optimization
mlx>=0.0.6
mlx-lm>=0.0.6

# Training & Evaluation
scipy>=1.11.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
wandb>=0.15.0

# Memory & Performance
memory-profiler>=0.61.0
psutil
GPUtil

# Text Processing
nltk>=3.8.0
rouge-score>=0.1.0
evaluate>=0.4.0

# Medical Evaluation
# medqa - removed as it's not available on PyPI
ragas>=0.0.18

# Utilities
structlog
pyyaml
python-dotenv
markupsafe~=2.0
packaging>=23.2,<24.0
rich>=13.0.0
typer>=0.12.1
click>=8.0.0

# Legacy LLM (backup)
llama-cpp-python

# Enhanced biomedical processing
biopython>=1.81
pubchempy>=1.0.4  # PubChem API access
datasets>=2.16.0

# Note: RDKit (chemical informatics) can be installed via pip or conda
# pip install rdkit (available for most platforms)
# For conda: conda install -c conda-forge rdkit
# We recommend conda installation for complex dependencies
# rdkit>=2024.3.1  # Uncomment after conda install or if pip install works

# Enhanced embeddings for biomedical domain
# sentence-transformers>=3.0.0  # Already included above
# bio-embeddings>=0.2.2  # Specialized protein embeddings - Install manually if needed

# Async HTTP Client
aiohttp>=3.9.0


